% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-list.R
\name{get_display_names}
\alias{get_display_names}
\title{Return display name for stats}
\usage{
get_display_names(stats)
}
\arguments{
\item{stats}{character vector of stats}
}
\value{
character vector of display names
}
\description{
Return display name for stats
}

