% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_star.R
\name{get_star}
\alias{get_star}
\title{Create the Siginicance Stars}
\usage{
get_star(pv, alpha, ...)
}
\arguments{
\item{pv}{vector of p values}

\item{alpha}{vector of significance levels}

\item{...}{not used
if \code{NA}, normal distribution is used.}
}
\description{
Based on the coefficient and standard errors and
significance levels, return character vector of stars.
}

