% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-list.R
\name{outreg_stat_list}
\alias{outreg_stat_list}
\title{List of Statictics Available on outreg}
\usage{
outreg_stat_list()
}
\value{
a data.frame that matches stat name and display name
}
\description{
Returns all available statistics on \code{\link{outreg}}.
Statistics names can be used for customizing the outputs, e.g.,
to choose stats to display or to choose stats to put starts.
}
\examples{
outreg_stat_list()
}

