% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outreg.R
\name{outreg}
\alias{outreg}
\title{Generate Regression Table}
\usage{
outreg(fitlist, digits = 3L, alpha = c(0.1, 0.05, 0.01), coef = TRUE,
  se = TRUE, pv = FALSE, tv = FALSE, bracket = c("se"),
  starred = c("coef"), robust = FALSE, small = TRUE)
}
\arguments{
\item{fitlist}{list of regression outcomes}

\item{digits}{number of digit to real numbers}

\item{alpha}{vector of significance levels to star}

\item{coef}{show coefficients}

\item{se}{show standard errors}

\item{pv}{show p values}

\item{tv}{show t values}

\item{bracket}{stats to be in brackets}

\item{starred}{stats to put stars on}

\item{robust}{if TRUE, robust standard error is used}

\item{small}{if TRUE, small sample parameter distribution is used}
}
\value{
regression table in data.frame format
}
\description{
Generate Regression Table
}
\examples{
fitlist <- list(lm(mpg ~ cyl, data = mtcars),
                lm(mpg ~ cyl + wt + hp, data = mtcars),
                lm(mpg ~ cyl + wt + hp + drat, data = mtcars))
outreg(fitlist)

# with custom regression names
outreg(setNames(fitlist, c('small', 'medium', 'large')))

# star on standard errors, instead of estimate
outreg(fitlist, starred = 'se')

# include other stats
outreg(fitlist, pv = TRUE, tv = TRUE, se = FALSE)

}

